ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pypdf